requirements.txt
setuptools
flask
langchain==0.3.26
pypdf==5.6.1 
python-dotenv==1.1.0
langchain-pinecone==0.2.8 
langchain-community==0.3.26
langchain-groq
-e .